chore: refactor projects structures
