+++ CardLibrary/Abstractions/IDeckShuffler.cs
@@ -1,0 +1,6 @@
+﻿namespace CardLibrary.Abstractions;
+
+public interface IDeckShuffler
+{
+    void Shuffle(ShuffleableCardDeck cardDeck);
+}
