+++ Colosseum/Abstractions/IExperiment.cs
@@ -1,0 +1,6 @@
+﻿namespace Colosseum.Abstractions;
+
+public interface IExperiment
+{
+    bool Do();
+}
