+++ StrategyLibrary/Interfaces/ICardPickStrategy.cs
@@ -1,0 +1,8 @@
+﻿using CardLibrary;
+
+namespace StrategyLibrary.Interfaces;
+
+public interface ICardPickStrategy
+{
+    int Choose(in CardDeck cardDeck);
+}
