+++ StrategyLibrary/StrategyLibrary.csproj
@@ -10,4 +10,6 @@
       <ProjectReference Include="..\CardLibrary\CardLibrary.csproj" />
     </ItemGroup>
 
+
+
 </Project>
